backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Scott Alan Turner <turner655+github@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.64"
psycopg2-binary = "^2.9.10"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
mysql-connector-python = "^9.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
